requirements.txt
arrow==1.3.0
blinker==1.7.0
buienradar==1.0.5
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
docopt==0.6.2
exceptiongroup==1.2.0
Flask==3.0.1
Flask-WTF==1.2.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
packaging==24.0
pluggy==1.4.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
pyparsing==3.1.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
types-python-dateutil==2.8.19.20240106
uritemplate==4.1.1
urllib3==2.2.0
vincenty==0.1.4
Werkzeug==3.0.1
WTForms==3.1.2
xmltodict==0.13.0
zipp==3.17.0
